pyproject.toml
[project]
name = "aps-automation-sdk"
version = "0.1.0"
description = "Python SDK for Autodesk Platform Services Automation API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=7.1.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[tool.setuptools]
packages = ["aps_automation_sdk"]
